# ui/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scanopy API",
    "description": "Network topology discovery and visualization API",
    "license": {
      "name": "Dual (AGPL3.0, Commercial License Available)"
    },
    "version": "0.12.5"
  },
  "paths": {
...
    "/api/discovery": {
      "get": {
        "tags": [
          "discoveries"
        ],
        "summary": "List all discoveries",
        "operationId": "list_discoveries",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "description": "Filter by network ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "daemon_id",
            "in": "query",
            "description": "Filter by daemon ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of discoveries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Discovery"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Create a new discovery",
        "operationId": "create_discovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Discovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "400": {
            "description": "Invalid subnet network"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/discovery/active-sessions": {
      "get": {
        "tags": [
          "discoveries"
        ],
        "summary": "Get active discovery sessions",
        "operationId": "get_active_sessions",
        "responses": {
          "200": {
            "description": "List of active discovery sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscoveryUpdatePayload"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/discovery/start-session": {
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Start a discovery session",
        "operationId": "start_session",
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryUpdatePayload"
                }
              }
            }
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/discovery/{id}": {
      "get": {
        "tags": [
          "discoveries"
        ],
        "summary": "Get discovery by ID",
        "operationId": "get_discovery_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Discovery ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovery found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "discoveries"
        ],
        "summary": "Update discovery",
        "operationId": "update_discovery",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Discovery ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Discovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "discoveries"
        ],
        "summary": "Delete discovery",
        "operationId": "delete_discovery",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Discovery ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovery deleted"
          },
          "404": {
            "description": "Discovery not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/discovery/{session_id}/cancel": {
      "post": {
        "tags": [
          "discoveries"
        ],
        "summary": "Cancel a discovery session",
        "operationId": "cancel_discovery",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovery session cancelled"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Discovery": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DiscoveryBase"
          },
          {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              }
            }
          }
        ]
      },
      "DiscoveryBase": {
        "type": "object",
        "required": [
          "discovery_type",
          "run_type",
          "name",
          "daemon_id",
          "network_id"
        ],
        "properties": {
          "daemon_id": {
            "type": "string",
            "format": "uuid"
          },
          "discovery_type": {
            "$ref": "#/components/schemas/DiscoveryType"
          },
          "name": {
            "type": "string"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "run_type": {
            "$ref": "#/components/schemas/RunType"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "DiscoveryPhase": {
        "type": "string",
        "enum": [
          "Pending",
          "Starting",
          "Started",
          "Scanning",
          "Complete",
          "Failed",
          "Cancelled"
        ]
      },
      "DiscoveryType": {
        "oneOf": [
          {
            "type": "object",
            "title": "SelfReport",
            "required": [
              "host_id",
              "type"
            ],
            "properties": {
              "host_id": {
                "type": "string",
                "format": "uuid"
              },
              "type": {
                "type": "string",
                "enum": [
                  "SelfReport"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "Network",
            "required": [
              "type"
            ],
            "properties": {
              "host_naming_fallback": {
                "$ref": "#/components/schemas/HostNamingFallback"
              },
              "subnet_ids": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "Network"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "Docker",
            "required": [
              "host_id",
              "type"
            ],
            "properties": {
              "host_id": {
                "type": "string",
                "format": "uuid"
              },
              "host_naming_fallback": {
                "$ref": "#/components/schemas/HostNamingFallback"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Docker"
                ]
              }
            }
          }
        ]
      },
      "DiscoveryUpdatePayload": {
        "type": "object",
        "description": "Progress update from daemon to server during discovery",
        "required": [
          "session_id",
          "daemon_id",
          "network_id",
          "phase",
          "discovery_type",
          "progress"
        ],
        "properties": {
          "daemon_id": {
            "type": "string",
            "format": "uuid"
          },
          "discovery_type": {
            "$ref": "#/components/schemas/DiscoveryType"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "network_id": {
            "type": "string",
            "format": "uuid"
          },
          "phase": {
            "$ref": "#/components/schemas/DiscoveryPhase"
          },
          "progress": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "session_id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "HostNamingFallback": {
        "type": "string",
        "enum": [
          "Ip",
          "BestService"
        ]
      },
...
      "RunType": {
        "oneOf": [
          {
            "type": "object",
            "title": "Scheduled",
            "required": [
              "cron_schedule",
              "enabled",
              "type"
            ],
            "properties": {
              "cron_schedule": {
                "type": "string"
              },
              "enabled": {
                "type": "boolean"
              },
              "last_run": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Scheduled"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "Historical",
            "required": [
              "results",
              "type"
            ],
            "properties": {
              "results": {
                "$ref": "#/components/schemas/DiscoveryUpdatePayload"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Historical"
                ]
              }
            }
          },
          {
            "type": "object",
            "title": "AdHoc",
            "required": [
              "type"
            ],
            "properties": {
              "last_run": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "type": {
                "type": "string",
                "enum": [
                  "AdHoc"
                ]
              }
            }
          }
        ],
        "description": "Note: DiscoveryUpdatePayload field is skipped from TS export as it's only used internally"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for daemon authentication. Generate keys in the web UI under Settings > API Keys."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}